learn_islam/graphql/users/types.py
from give_core.users.models import User
from graphene_django.types import DjangoObjectType


class UserType(DjangoObjectType):
    class Meta:
        model = User
        exclude_fields = ('password',)

